fix(tests): compare apsp hop counts in their signed type
Convert the bounded loop hop count to the source path integer type so the S1a driver compiles without changing its fixture or exact path-weight criterion.

+++ tests/upstream/examples/apsp.rs
@@ -218,7 +218,7 @@ fn sparse_path_doubling<'c, 'r>(
         let old_paths = paths.clone();
         let paths_i = old_paths
             .clone()
-            .into_sparse(|path| path.hops == hops);
+            .into_sparse(|path| path.hops == hops as i32);
         let global_nnz = canonical_nnz(&paths_i);
         let selected = cache
             .prepare(
